API

TMatrix.AbstractScatterer

TMatrix.Bicone

TMatrix.Bicone

TMatrix.Capsule

TMatrix.Capsule

TMatrix.Chebyshev

TMatrix.Cylinder

TMatrix.Cylinder

TMatrix.ScattererInfo

TMatrix.ScattererInfo

TMatrix.ScattererInfo

TMatrix.Spheroid

TMatrix.Spheroid

TMatrix.calc_SZ

TMatrix.calc_Z

TMatrix.calc_amplitude

TMatrix.calc_expansion_coefficients

TMatrix.calc_phase

TMatrix.calc_r

TMatrix.calc_r!

TMatrix.calc_r!

TMatrix.calc_scattering_matrix

TMatrix.calc_scattering_matrix

TMatrix.calc_tmatrix!

TMatrix.calc_tmatrix!

TMatrix.hovenr

TMatrix.set_arb_approx_inv

TMatrix.set_default_ncap

TMatrix.set_default_ngcap

TMatrix.set_default_ngcheb

TMatrix.sphericalbesselj!

TMatrix.tmatrix_routine_mishchenko

TMatrix.tmatrix_routine_mishchenko_nmaxonly

TMatrix.AbstractScatterer — Type

Abstract type for all scatterers.

TMatrix.Bicone — Type

Construct a bicone.

TMatrix.Bicone — Type

A bicone scatterer.

Attributes:

m: The complex refractive index.

r: The radius of the cone.

h: The height of one cone.

λ: The wavelength of the incident wave.

info: The accompanied information.

TMatrix.Capsule — Type

Construct a capsule.

TMatrix.Capsule — Type

A capsule scatterer, which is a cylinder with half-spherical caps on both ends.

Attributes:

m: The complex refractive index.

r: The radius of the cylinder base and the hemisphere.

h: The height of the cylinder.

λ: The wavelength of the incident wave.

info: The accompanied information.

TMatrix.Chebyshev — Type

A Chebyshev scatterer defined by

$r(\theta, \phi)=r_0(1+\varepsilon T_n(\cos\theta))$

in which $T_n(\cos\theta)=\cos n\theta$.

Attributes:

m: The complex refractive index.

r₀: The radius of the base sphere.

ε: The deformation parameter, which satisfies $-1\le\varepsilon<1$.

n: The degree of the Chebyshev polynomial.

λ: The wavelength of the incident wave.

info: The accompanied information.

TMatrix.Cylinder — Type

Construct a cylinder. λ and m must be provided, and the size parameters are considered according to the following priority:

r and h

r and r_to_h

h and r_to_h

rev and r_to_h

rea and r_to_h

rmax and r_to_h

If none of the above is hit, an ArgumentError will be thrown.

TMatrix.Cylinder — Type

A cylinder scatterer.

Attributes:

m: The complex refractive index.

r: The radius of the cylinder base.

h: The height of the cylinder.

λ: The wavelength of the incident wave.

info: The accompanied information.

TMatrix.ScattererInfo — Type

Accompanied information of a scatterer.

TMatrix.ScattererInfo — Method

Constructor of ScattererInfo for general data types. Space is pre-allocated to reduce allocations.

TMatrix.ScattererInfo — Method

Constructor of ScattererInfo for Arb. Pre-assignment is not used, since SubArray does not work harmoniously with ArbMatrix.

TMatrix.Spheroid — Type

Construct a spheroid. λ and m must be provided, and the size parameters are considered according to the following priority:

a and c

a and a_to_c

c and a_to_c

rev and a_to_c

rea and a_to_c

rmax and a_to_c

If none of the above is hit, an ArgumentError will be thrown.

TMatrix.Spheroid — Type

A spheroid scatterer.

Attributes:

m: The complex refractive index.

a: Length of the semi-major axis.

c: Length of the semi-minor axis.

λ: The wavelength of the incident wave.

info: The accompanied information.

TMatrix.calc_SZ — Method

calc_SZ(scatterer::AbstractScatterer{T}, α::T, β::T, ϑ_i::T, ϑ_s::T, φ_i::T, φ_s::T, TT::Vector{<:AbstractMatrix}) where {T<:Real}

Calculate the S matrix and the Z matrix sequentially.

TMatrix.calc_Z — Function

calc_Z(S::AbstractMatrix)

Alias of calc_phase

TMatrix.calc_amplitude — Method

calc_amplitude(scatterer::AbstractScatterer{T}, α::T, β::T, ϑ_i::T, ϑ_s::T, φ_i::T, φ_s::T, TT::Vector{<:AbstractMatrix}) where {T<:Real}

Calculate the amplitude matrix and the phase matrix, given the scatterer and the geometry of the incident and the scattered beam. Use pre-computed T-Matrix when possible.

Parameters:

scatterer: The scatterer.

α, β: The Euler angle.

ϑ_i: The zenith angle of the incident beam.

ϑ_s: The zenith angle of the scattered beam.

φ_i: The azimuth angle of the indicent beam.

φ_s: The azimuth angle of the scatterer beam.

TT: The pre-computed T-Matrix of the scatterer.

All the angles here are input in degrees.

TMatrix.calc_expansion_coefficients — Method

calc_expansion_coefficients(TT::Vector{<:AbstractMatrix}, Csca::Real, λ::Real)

Calculate the expansion coefficients from a given T-Matrix.

Parameters:

TT: The precalculated T-Matrix of a scatterer.

Csca: The scattering cross setction.

λ: The wavelength.

Csca and λ should have the same unit of length. E.g., if λ is in μm, Csca should be in μm².

TMatrix.calc_phase — Method

calc_phase(S::AbstractMatrix)

Calculate the phase matrix using the given amplitude matrix $\mathbf{S}$.

Parameters:

S, the precalculated amplitude matrix.

TMatrix.calc_r! — Method

calc_r!(scatterer::Capsule{T}, ngauss::Int64, x::AbstractArray{T}, w::AbstractArray{T}, r::AbstractArray{T}, dr::AbstractArray{T}) where {T<:Real}

Calculate $r(\theta)$ and $\frac{\mathrm{d}r}{\mathrm{d}\theta}$ at ngauss points for a given capsule, in place.

TMatrix.calc_r! — Method

calc_r!(scatterer::AbstractScatterer{T}, ngauss::Int64, x::AbstractArray{T}, w::AbstractArray{T}, r::AbstractArray{T}, dr::AbstractArray{T}) where {T<:Real}

Calculate $r(\theta)$ and $\frac{\mathrm{d}r}{\mathrm{d}\theta}$ at ngauss points for a given scatterer, in place.

TMatrix.calc_r — Method

calc_r(scatterer::Scatterer, ngauss::Int64)

Calculate $r(\theta)$ and $\frac{\mathrm{d}r}{\mathrm{d}\theta}$ at ngauss points for a given scatterer.

TMatrix.calc_scattering_matrix — Method

calc_scattering_matrix(scatterer, TT, Nθ)

Calculate the scatterering matrix elements from the given scatterer and precalculated T-Matrix.

Parameters:

scatterer: The scatterer.

TT: The T-Matrix.

Nθ: Number of θ intervals (so the result will have Nθ + 1 rows).

TMatrix.calc_scattering_matrix — Method

calc_scattering_matrix(α₁, α₂, α₃, α₄, β₁, β₂, θ)

Calculate the scatterering matrix elements from the given expansion coefficients.

Parameters:

α₁, α₂, α₃, α₄, β₁, β₂: The precalculated expansion coefficients.

θ: The scattering angle in degrees.

TMatrix.calc_tmatrix! — Method

calc_tmatrix!(scatterer::AbstractScatterer{T}) where {T<:Real}

Calculate the T-Matrix of the scatterer with default settings:

Use tmatrix_routine_mishchenko to generate the routine function.

Use ddelta = 0.001 and ndgs = 4.

Parameters:

scatterer: The scatterer.

TMatrix.calc_tmatrix! — Method

calc_tmatrix(scatterer::AbstractScatterer{T}, ngstart::Int64, nstart::Int64, routine::Function) where {T<:Real}

Calculate the T-Matrix of the scatterer.

Parameters:

scatterer: The scatterer.

ngstart: The starting point of ngauss.

nstart: The starting point of nmax.

routine: The iteration routine function generated by a routine generator, internally or customly implemented.

TMatrix.hovenr — Method

hovenr(α₁, α₂, α₃, α₄, β₁, β₂)

Validate the expansion coefficients with the test of Van der Mee & Hovenier.

Note that the index of the coefficients should start from $0$.

TMatrix.set_arb_approx_inv — Function

set_arb_approx_inv(t::Bool = true)

Turn on/off Arblib's approx_inv for ArbMatrix and AcbMatrix.

TMatrix.set_default_ncap — Function

set_default_ncap(ncap::Int64 = 100)

Change the default ncap value.

TMatrix.set_default_ngcap — Function

set_default_ngcap(ngcap::Int64 = 500)

Change the default ngcap value.

TMatrix.set_default_ngcheb — Function

set_default_ngcheb(ngcheb::Int64 = 60)

Change the default ngcheb value.

TMatrix.sphericalbesselj! — Method

sphericalbesselj!(x::T, nmax::Int64, nnmax1::Int64, y::AbstractArray{T}, u::AbstractArray{T}, z::AbstractArray{T}) where {T <: Number}

Calculate spherical Bessel function $j_n(x)$ and $\frac{1}{x}\frac{\mathrm{d}}{\mathrm{d}x}[xj_n(x)]$ in place.

TMatrix.tmatrix_routine_mishchenko — Method

tmatrix_routine_mishchenko(scatterer::AbstractScatterer{T}, ddelta::T, ndgs::Int64) where {T<:Real}

Generate the routine function which follows the iteration strategy used by M. Mishchenko.

Parameters:

scatterer: The scatterer.

ddelta: The convergence threshold.

ndgs: Determines the initial value of ngauss as ndgs * nmax. The initial value of nmax is determined by the formula $\max(4, \lfloor kr + 4.05 * \sqrt[3]{kr}\rfloor)$.

nstart: Optional for manual setting of the initial value of nmax.

ngstart: Optional for manual setting of the initial value of ngauss. This parameter takes effect only if nstart≠0.

TMatrix.tmatrix_routine_mishchenko_nmaxonly — Method

tmatrix_routine_mishchenko_nmaxonly(scatterer::AbstractScatterer{T}, ddelta::T, ndgs::Int64) where {T<:Real}

Generate the routine function which generally follows the iteration strategy used by M. Mishchenko, but does not modify the value of ngauss.

Parameters:

scatterer: The scatterer.

ddelta: The convergence threshold.

ndgs: Determines the initial value of ngauss as ndgs * nmax. The initial value of nmax is determined by the formula $\max(4, \lfloor kr + 4.05 * \sqrt[3]{kr}\rfloor)$.

nstart: Optional for manual setting of the initial value of nmax.

ngstart: Optional for manual setting of the initial value of ngauss. This parameter takes effect only if nstart≠0.